core - create Confirm code screen(need fix serilization in desktop)

+++ common/auth/api/src/commonMain/kotlin/models/ConfirmCode.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfirmCode(
-    val code: Int
+    val code: Long
 )
